Update GoogleAuthenticator.cfc
This version contains code to allow and report time drift. This is required if you deal with hardware based tokens. The longer since last used, the more grace that needs to be added to compensate for time drift. Maintaining a drift value of last used per device, and the date of last used, allows for less grace to be needed. Clear as mud, huh? There is an RFC on the topic (see RFC 6238 , section 6).